Some API endpoints now output an array of expected/missing data.
The API endpoints for users, falls, and diary entries now output an array of data that the endpoint expects with every request, and what was missing from the request that was made.

+++ web/public/api/falls/delete.php
@@ -8,6 +8,9 @@
 
 		$api_key = isset($_GET['key']) ? $_GET['key'] : die(json_encode(array('message' => 'No API key provided.')));
 
+		$expected = [];
+		$missing = [];
+
 		$database = new Database();
 		$db = $database->connect($api_key);
 
@@ -15,9 +18,13 @@
 		
 		$input = json_decode(file_get_contents('php://input'), true);
 
-		$fall->fallID = !empty($input['fallID']) ? $input['fallID'] : die();
+		$fall->fallID = !empty($input['fallID']) ? $input['fallID'] : array_push($missing, 'fallID');
 
-		$fall->delete();
+		if(empty($missing)) {
+			$fall->delete();
+		} else {
+			die(json_encode(array('expected' => $expected, 'missing' => $missing), JSON_PRETTY_PRINT));
+		}
 	} else {
 		echo json_encode(array('message' => 'Wrong HTTP request method. Use DELETE instead.'));
 	}
