+++ web/public/api/users/delete.php
@@ -8,6 +8,9 @@
 
 		$api_key = isset($_GET['key']) ? $_GET['key'] : die(json_encode(array('message' => 'No API key provided.')));
 
+		$expected = [];
+		$missing = [];
+
 		$database = new Database();
 		$db = $database->connect($api_key);
 
@@ -15,9 +18,13 @@
 		
 		$input = json_decode(file_get_contents('php://input'), true);
 
-		$user->patientID = !empty($input['patientID']) ? $input['patientID'] : die();
+		$user->patientID = !empty($input['patientID']) ? $input['patientID'] : array_push($missing, 'patientID');
 
-		$user->delete();
+		if(empty($missing)) {
+			$user->delete();
+		} else {
+			die(json_encode(array('expected' => $expected, 'missing' => $missing), JSON_PRETTY_PRINT));
+		}
 	} else {
 		echo json_encode(array('message' => 'Wrong HTTP request method. Use DELETE instead.'));
 	}
